--- a/benchmarks/LetsTrace.Benchmarks/TagBenchmark.cs
+++ b/benchmarks/LetsTrace.Benchmarks/TagBenchmark.cs
@@ -2,6 +2,7 @@ using BenchmarkDotNet.Attributes;
 
 namespace LetsTrace.Benchmarks
 {
+    [MemoryDiagnoser]
     public class TagBenchmark
     {
         private readonly Tracer _tracer;