--- a/src/LetsTrace/Samplers/ISampler.cs
+++ b/src/LetsTrace/Samplers/ISampler.cs
@@ -10,6 +10,6 @@ namespace LetsTrace.Samplers
         // can be used to identify the type of sampling that was applied to
         // the root span. Most simple samplers would return two tags,
         // sampler.type and sampler.param, similar to those used in the Configuration
-        (bool Sampled, IDictionary<string, Field> Tags) IsSampled(TraceId id, string operation);
+        (bool Sampled, Dictionary<string, object> Tags) IsSampled(TraceId id, string operation);
     }
 }