--- a/src/LetsTrace/Samplers/PerOperationSampler.cs
+++ b/src/LetsTrace/Samplers/PerOperationSampler.cs
@@ -88,7 +88,7 @@ namespace LetsTrace.Samplers
             return isUpdated;
         }
 
-        public (bool Sampled, IDictionary<string, Field> Tags) IsSampled(TraceId id, string operation)
+        public (bool Sampled, Dictionary<string, object> Tags) IsSampled(TraceId id, string operation)
         {
             var operationKey = operation.ToLower();