… src/LetsTrace/Field.cs
… src/LetsTrace/FieldExtensions.cs
… test/LetsTrace.Tests/FieldTests.cs
… test/LetsTrace.Tests/SpanBuilderTests.cs
… test/LetsTrace.Tests/SpanTests.cs