As discussed briefly in https://github.com/opentracing/opentracing-csharp/issues/69 it's not really necessary to have this separate `Field` class as you can also just switch on the type. Not using a wrapper-type also results in better performance and less memory pressure.

This PR would therefore remove the `Field` type and bring the logic back on par with `jaeger-client-java`. 

One thing I've not handled yet though is the serialization/handling of non-valid types. While these values have been converted to string by the `Field` type before, they must now be handled by whomever enumerates them. This is the same behavior of jaeger-client-java though.

#### Before:
``` ini
BenchmarkDotNet=v0.10.13, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.309)
Intel Core i5-4300U CPU 1.90GHz (Haswell), 1 CPU, 4 logical cores and 2 physical cores
Frequency=2435766 Hz, Resolution=410.5485 ns, Timer=TSC
.NET Core SDK=2.1.300-preview1-008174
  [Host]     : .NET Core 2.0.6 (CoreCLR 4.6.26212.01, CoreFX 4.6.26212.01), 64bit RyuJIT
  DefaultJob : .NET Core 2.0.6 (CoreCLR 4.6.26212.01, CoreFX 4.6.26212.01), 64bit RyuJIT
```
|                               Method |       Mean |    Error |   StdDev |  Gen 0 | Allocated |
|------------------------------------- |-----------:|---------:|---------:|-------:|----------:|
|                                NoTag |   860.3 ns | 15.34 ns | 14.35 ns | 0.5178 |     816 B |
|                OneTag_on_SpanBuilder |   915.0 ns | 12.67 ns | 11.85 ns | 0.5493 |     864 B |
|               TwoTags_on_SpanBuilder | 1,089.6 ns | 21.14 ns | 20.76 ns | 0.7362 |    1160 B |
|                       OneTag_on_Span |   912.9 ns | 17.38 ns | 15.41 ns | 0.5493 |     864 B |
|                      TwoTags_on_Span | 1,089.5 ns | 26.50 ns | 27.21 ns | 0.7362 |    1160 B |
| OneTag_on_SpanBuilder_OneTag_on_Span | 1,092.3 ns | 19.85 ns | 29.72 ns | 0.7362 |    1160 B |

#### After:
``` ini
BenchmarkDotNet=v0.10.13, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.309)
Intel Core i5-4300U CPU 1.90GHz (Haswell), 1 CPU, 4 logical cores and 2 physical cores
Frequency=2435766 Hz, Resolution=410.5485 ns, Timer=TSC
.NET Core SDK=2.1.300-preview1-008174
  [Host]     : .NET Core 2.0.6 (CoreCLR 4.6.26212.01, CoreFX 4.6.26212.01), 64bit RyuJIT
  DefaultJob : .NET Core 2.0.6 (CoreCLR 4.6.26212.01, CoreFX 4.6.26212.01), 64bit RyuJIT
```
|                               Method |     Mean |     Error |    StdDev |  Gen 0 | Allocated |
|------------------------------------- |---------:|----------:|----------:|-------:|----------:|
|                                NoTag | 796.4 ns |  9.945 ns |  9.303 ns | 0.4826 |     760 B |
|                OneTag_on_SpanBuilder | 816.4 ns |  7.831 ns |  6.942 ns | 0.4826 |     760 B |
|               TwoTags_on_SpanBuilder | 953.6 ns | 11.635 ns | 10.314 ns | 0.6399 |    1008 B |
|                       OneTag_on_Span | 821.9 ns |  9.170 ns |  8.577 ns | 0.4826 |     760 B |
|                      TwoTags_on_Span | 954.0 ns | 19.062 ns | 18.722 ns | 0.6390 |    1008 B |
| OneTag_on_SpanBuilder_OneTag_on_Span | 958.6 ns | 20.628 ns | 19.295 ns | 0.6390 |    1008 B |

PS: My PRs today probably broke a lot in LetsTrace.Jaeger & LetsTrace.Zipkin - I'll look at them as well of course.