@grounded042 2018-03-26
Thanks @cwe1ss. This looks good. I do agree that dropping `Field` makes sense - especially with the performance considerations. 

As far as breaking LetsTrace.Jaeger, there's actually a PR (#10) for merging LetsTrace.Jaeger into here so don't worry about looking into that repo as I can look at taking care of the changes from PRs.